requirements.txt
requests
google-cloud-secret-manager==2.20.2
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pytz
functions-framework==3.*
Flask==3.0.3
psycopg2-binary==2.9.9
pydantic==2.5.3
sqlalchemy==2.0.23